[*] Rewritten pinning logic. Now public key hash can be added dynamically. [+] Added sepate classes for Tabbar and NavigationBar [+] Almost completely added detail view for translation [*] Rewritten image cache to swift [f] Fixed crashes when trying to download old localizations [f] No app icon is now creates programmatically [f] Fixed crash in posix_spawn when trying to execute invalid command

+++ RJTranslate-App/Views/RJTHud.m
@@ -291,7 +291,7 @@ static CGFloat const RJTHudContentSize = 140.0f;
     
     CGFloat height = progressViewHeight + MAX(textSize.height, minLabelHeight) + MAX(detailTextSize.height, minLabelHeight);
     if (height > RJTHudContentSize)
-        height += 32.0;
+        height += (CGFloat)32.0;
     
     self.heightConstraint.constant = height;
     self.textLabelHeightConstraint.constant = textSize.height;
